Move cliinputjson tests from integration to functional
